requirements.txt
boto3==1.26.69
botocore==1.29.69
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
dnspython==2.3.0
idna==3.4
jmespath==1.0.1
pycparser==2.21
pymongo==4.3.3
pymongo-auth-aws==1.1.0
pymongocrypt==1.5.1
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
